S3-327-90%-completed
The highlight, filter, display all and dropdown menus have all been integrated. In addition to that, the search function which is not in this task are also completed and integrate with display table data. The 10% is left is fixing the small element of how the data should be display and little bugs.