S3-327-100%-completed
S3-327-All the bugs are all fixed. The filtering and highlighting mode are updated where it is allowing the the user to search the firstname of the student without losing the colour of flag. I have also fixed the bug where it table displaying supposed to return unknownStudents object when they cannot find the match while the user using highlight or filter mode.